+++ components/content/slides/default.js
